move css logic into NoJSButton
